feat+refactor: ban the account associate with ip; move moderation code into seperate endpoints
